refactor: переименовать компонент
