+++ src/entities/clinic/index.ts
@@ -1,4 +1,4 @@
 import ClinicList from './ui/ClinicList/ClinicList';
-import { FullCardClinic } from './ui/FullCardClinic/';
+import { FullCardClinic } from './ui/FullCardClinic/FullCardClinic';
 
 export { ClinicList, FullCardClinic };
