+++ src/entities/clinic/ui/FullCardClinic/FullCardClinic.scss
